packages/slack-bot/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
contourpy==1.0.7
cycler==0.11.0
deprecation==2.1.0
Flask==2.2.3
fonttools==4.39.0
frozenlist==1.3.3
functions-framework==3.3.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.1
plotly==5.13.1
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
slack-bolt==1.16.4
slack-sdk==3.20.2
tenacity==8.2.2
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
watchdog==2.3.1
Werkzeug==2.2.3
yarl==1.8.2
